remove constants files
