+++ src/components/src/ProductDisplayItemDetails/productdisplayitem.details.tsx
@@ -28,12 +28,68 @@ import { useRequisitionListCountDispatch } from '../requisition-list-count-conte
 import { ProductDisplayDetailsProps, ProductDisplayItemMainState } from './productdisplayitem.details.d';
 import SocialNetworkSharing from '../SocialNetworkSharing/socialNetworkSharing';
 import QuantitySelector from '../QuantitySelector/quantitySelector';
-import { ZOOM } from './productdisplayitem.details.constants';
 import './productdisplayitem.details.less';
 
 let intl = { get: str => str };
 
 let Config: IEpConfig | any = {};
+
+export const ZOOM : string[] = [
+  'availability',
+  'addtocartform',
+  'addtocartforms:element:addtocartaction',
+  'addtocartforms:element:target:descriptor',
+  'addtowishlistform',
+  'price',
+  'rate',
+  'definition',
+  'definition:assets:element',
+  'definition:options:element',
+  'definition:options:element:value',
+  'definition:options:element:selector:choice',
+  'definition:options:element:selector:chosen',
+  'definition:options:element:selector:choice:description',
+  'definition:options:element:selector:chosen:description',
+  'definition:options:element:selector:choice:selector',
+  'definition:options:element:selector:chosen:selector',
+  'definition:options:element:selector:choice:selectaction',
+  'definition:options:element:selector:chosen:selectaction',
+  'definition:components',
+  'definition:components:element',
+  'definition:components:element:code',
+  'definition:components:element:standaloneitem',
+  'definition:components:element:standaloneitem:code',
+  'definition:components:element:standaloneitem:definition',
+  'definition:components:element:standaloneitem:availability',
+  'recommendations',
+  'recommendations:crosssell',
+  'recommendations:recommendation',
+  'recommendations:replacement',
+  'recommendations:upsell',
+  'recommendations:warranty',
+  'recommendations:crosssell:element:code',
+  'recommendations:recommendation:element:code',
+  'recommendations:replacement:element:code',
+  'recommendations:upsell:element:code',
+  'recommendations:warranty:element:code',
+  'recommendations:crosssell:element:definition',
+  'recommendations:recommendation:element:definition',
+  'recommendations:replacement:element:definition',
+  'recommendations:upsell:element:definition',
+  'recommendations:warranty:element:definition',
+  'recommendations:crosssell:element:price',
+  'recommendations:recommendation:element:price',
+  'recommendations:replacement:element:price',
+  'recommendations:upsell:element:price',
+  'recommendations:warranty:element:price',
+  'recommendations:crosssell:element:availability',
+  'recommendations:recommendation:element:availability',
+  'recommendations:replacement:element:availability',
+  'recommendations:upsell:element:availability',
+  'recommendations:warranty:element:availability',
+  'code',
+];
+
 
 class ProductDisplayItemDetails extends Component<ProductDisplayDetailsProps, ProductDisplayItemMainState> {
   constructor(props) {
